deleted old page (template-two) & removed top toolbar
